requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.19.0
accelerate>=0.21.0

# Image processing
Pillow>=9.5.0
numpy>=1.24.0

# Utilities
python-dotenv>=0.19.0
safetensors>=0.3.1
requests>=2.31.0
tqdm>=4.65.0
matplotlib>=3.7.0

# Optional - for better performance
triton>=2.0.0; platform_system=="Linux"
xformers>=0.0.20; platform_system!="Darwin"